fix(cairo): cairo.region not accepted as argument
